requirements.txt
flask==2.3.3
tensorflow==2.13.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
matplotlib==3.7.2
lime==0.2.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
requests==2.31.0
eli5==0.13.0
pandas==2.0.3